Added booking by distance, added Consts class containing constants like price per time/distance and base price, changed database total price field from int to double

+++ MooveTeq-Booking/Data/Booking.cs
@@ -12,7 +12,7 @@ namespace MooveTeqBooking.Data {
 
         public DateTime? EndTime { get; set; }
 
-        public int? TripDistance { get; set; }
+        public double? TripDistance { get; set; }
 
         public double TotalTripCost { get; set; }
